backend/requirements.txt
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
attrs==25.4.0
audioop-lts==0.2.2
audioread==3.0.1
backoff==2.2.1
bcrypt==5.0.0
blockbuster==1.5.25
build==1.3.0
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.3
chromadb==1.1.1
click==8.3.0
cloudpickle==3.1.1
coloredlogs==15.0.1
crcmod==1.7
cryptography==44.0.3
datasets==4.2.0
decorator==5.2.1
dill==0.4.0
distro==1.9.0
dnspython==2.8.0
durationpy==0.10
editdistance==0.8.1
einops==0.8.1
email-validator==2.3.0
fastapi==0.118.2
fastapi-cli==0.0.13
fastapi-cloud-cli==0.3.1
filelock==3.20.0
flatbuffers==25.9.23
forbiddenfruit==0.1.4
frozenlist==1.8.0
fsspec==2025.9.0
funasr==1.2.7
google-auth==2.41.1
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-tools==1.75.1
h11==0.16.0
hf-xet==1.1.10
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface-hub==0.35.3
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
iniconfig==2.1.0
inquirerpy==0.3.4
itsdangerous==2.2.0
jaconv==0.4.0
jamo==0.4.1
jieba==0.42.1
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jsonschema_rs==0.29.1
kaldiio==2.18.1
kubernetes==34.1.0
langchain-core==0.3.79
langgraph==0.6.10
langgraph-api==0.4.39
langgraph-checkpoint==2.1.2
langgraph-cli==0.4.3
langgraph-prebuilt==0.6.4
langgraph-runtime-inmem==0.14.1
langgraph-sdk==0.2.9
langsmith==0.4.34
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.45.1
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mmh3==5.2.0
modelscope==1.30.0
mpmath==1.3.0
msgpack==1.1.2
multidict==6.7.0
multiprocess==0.70.16
networkx==3.5
numba==0.62.1
numpy==1.26.4
oauthlib==3.3.1
omegaconf==2.3.0
onnxruntime==1.23.1
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
orjson==3.11.3
ormsgpack==1.11.0
oss2==2.19.1
overrides==7.7.0
packaging==25.0
pandas==2.3.3
pfzy==0.3.4
pillow==11.3.0
platformdirs==4.5.0
pluggy==1.6.0
pooch==1.8.2
posthog==5.4.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.32.1
psutil==7.1.0
psycopg2-binary==2.9.11
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybase64==1.4.2
pycparser==2.23
pycryptodome==3.23.0
pydantic==2.12.0
pydantic-extra-types==2.10.6
pydantic-settings==2.11.0
pydantic_core==2.41.1
Pygments==2.19.2
PyJWT==2.10.1
pynndescent==0.5.13
PyPika==0.48.9
pyproject_hooks==1.2.0
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.20
pytorch-wpe==0.0.1
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==14.2.0
rich-toolkit==0.15.1
rignore==0.7.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
sentencepiece==0.2.1
sentry-sdk==2.41.0
setuptools==80.9.0
shellingham==1.5.4
sherpa-onnx==1.12.14
sherpa-onnx-core==1.12.14
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
soxr==1.0.0
SQLAlchemy==2.0.43
sse-starlette==2.1.3
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
starlette==0.48.0
structlog==25.4.0
sympy==1.14.0
tenacity==9.1.2
tensorboardX==2.6.4
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.8.0
torch-complex==0.4.4
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.57.0
truststore==0.10.4
typer==0.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
umap-learn==0.5.9.post2
urllib3==2.3.0
uvicorn==0.37.0
uvloop==0.21.0
watchfiles==1.1.0
wcwidth==0.2.14
websocket-client==1.9.0
websockets==15.0.1
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
zstandard==0.25.0
